camera: cache TET value when camera device is closed
Cache the last Gcam TET value and HDR ratio when we close the camera
device. The next time the camera device is opened, the cached settings
will be used as the initial parameters of the AE loop. This speeds up
the AE convergence when the scene does not differ too much between the
two consecutive camera sessions.

BUG=b:210364439
TEST=Manually on device with various scenes.

Change-Id: Ib4ab0e59e357b4d3d89c9eea6c297fc1aae3f5d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3337022
Tested-by: Ricky Liang <jcliang@chromium.org>
Reviewed-by: Wei Lee <wtlee@chromium.org>
Commit-Queue: Ricky Liang <jcliang@chromium.org>