shill: vpn: Add service properties in IKEv2Driver
This patch adds properties in IKEv2Driver and the code to map them to
IPsecConnection::Config. Note that we use the EapCredentials class to
handle the EAP-related fields (currently only MSCHAPv2), we keep an
instance of EapCredentials inside the base VPNDriver class instead of
the IKEv2Driver class, so that we don't need add several hooks and
implement them in IKEv2Driver class.

BUG=b:210064468
TEST=unit_tests

Change-Id: I6cfd03b937582e5162e91efb25685c31d6d85c16
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3341012
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>