camera: adjust framing ROI based on active stream dimension
When the active camera stream is cropped from the active sensor array to
a letterbox or pillarbox, we should adjust the normalized rectangle
coordinates of the ROI accordingly. Otherwise, for example, the face in
the ROI will look bigger on a 16:9 stream than on a 4:3 stream on the
same device if the device has a 4:3 sensor.

BUG=b:204748009
TEST=Manually on Redrix with 4:3 and 16:9 streams and verify that the
face size is the same in both stream configurations.

Change-Id: I217d01f6985dd2f07153b417d5bcc6269f626607
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3341015
Tested-by: Ricky Liang <jcliang@chromium.org>
Reviewed-by: Ren-Pei Zeng <kamesan@chromium.org>
Commit-Queue: Ricky Liang <jcliang@chromium.org>