system_api: Add new handlers for ArcVmCompleteBoot
The current handler naming of MakeRtVCpu will be transitioned to
ArcVmCompleteBoot to better represent the intent of the event.

These changes add the new defiinitions and handlers for this
API. Future changes will remove the old handlers once Chrome usage
is updated to ArcVmCompleteBoot.

BUG=b:199825292
TEST=tast run dut arc.PlayStore.vm

Change-Id: I55acdee8d338e0c3f22630137f0c1789b1cb790b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3343588
Reviewed-by: Charles William Dick <cwd@google.com>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Reviewed-by: Abhishek Bhardwaj <abhishekbh@chromium.org>
Tested-by: Kameron Lutes <kalutes@chromium.org>
Commit-Queue: Kameron Lutes <kalutes@chromium.org>