fusebox: define template function to call the fusebox server
Calling the server is a frequent operation. To reduce boilerplate
some [1], define a C++ template for calling server methods.

Move the GetFuseBoxServerMethodCall and the template together, to
connect the idea of 1) getting a server D-BUS method_call, and 2)
calling that D-BUS method_call.

[1] Only hides dbus::ObjectProxy::TIMEOUT_USE_DEFAULT which D-BUS
defines as "infinite". In Chrome / ChromeOS D-BUS implementation,
the default timeout is 30 seconds.

BUG=chromium:1244007
TEST=emerge --board=${BOARD} fusebox && deploy

Change-Id: I408ccc534678cf8cf7002e34438726b042977656
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3350276
Reviewed-by: François Degros <fdegros@chromium.org>
Commit-Queue: Noel Gordon <noel@chromium.org>
Tested-by: Noel Gordon <noel@chromium.org>