fusebox: add FUSE Open and Release ops
Add FUSE op Open: an optional FUSE op, used if a FUSE user-space
session wants to allocate a file handle for open(2).

This implementation will allocate file handles. File handles are
maintained in the client and returned to Kernel FUSE, which uses
them as file descriptors (fd) for POSIX API callers.

How do we know that the given ino is a file? Kernel FUSE sends a
a GetAttr or Lookup op before an Open op, to check the ino perms
and entry mode (stat) is a file that can be opened.

The Open request is sent to server, which returns an optional fd
that the client stores in the file handle.

Add FUSE op Release for close(2) that undoes the work of FUSE op
Open. Close the file handle (and its underlying fd).

Performance: the server open request call to chrome::storage API
FileSystemOperationRunner::OpenFile (for an MTP device Pixel 3XL
phone) takes 290 milliseconds (average) on my nami DUT.

BUG=chromium:1244007
TEST=emerge --board=nami fusebox && deploy

Change-Id: I80b4b3b90a58758363527dadf75a888f3337b97b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3347497
Reviewed-by: François Degros <fdegros@chromium.org>
Commit-Queue: Noel Gordon <noel@chromium.org>
Tested-by: Noel Gordon <noel@chromium.org>