fusebox: add FUSE Read op [part 1 of 2] file descriptor reader
FUSE Open op opened a file handle for the ino and also provided an
optional file descriptor (fd) that was stored in the file handle.

Add FUSE op Read, for read(2) and related POSIX reading API. POSIX
advises that the read |size| should be <= SSIZE_MAX. Fail the FUSE
Read op with EINVAL if |size| > SSIZE_MAX.

The read |size| is limited by the Kernel in practice: 131072 bytes
at most on my nami DUT.

Part 1 case: the file descriptor was provided and is not -1, which
is the invalid fd value. Use it to pread(2) the file data locally,
and return the data buffer.

Add TODO for part 2 case: reading the file's data from the server,
which will be the subject of the next patch.

Performance: 358 MB/s for the read operations (3MB test file). But
as noted in CL:3347497, the FUSE Open op takes 290 milliseconds on
the server. That time dominates the read file flow. If included in
the overall time, the effective file read rate is only 9 MB/s [1].

[1] These numbers for a Pixel 3 XL phone MTP device mounted in the
FuseBoxClient brillo::daemon. Android Version 11.

BUG=chromium:1244007
TEST=emerge --board=nami fusebox && deploy

Change-Id: I71f923ad5ec4f06cba47218ecf475a660d7dbc18
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3353065
Reviewed-by: François Degros <fdegros@chromium.org>
Commit-Queue: Noel Gordon <noel@chromium.org>
Tested-by: Noel Gordon <noel@chromium.org>