fusebox: add FUSE Read op [part 2 of 2] server file reader
FUSE Open op opened a file handle for the ino and also provided an
optional file descriptor (fd) that was stored in the file handle.

If no fd was provided, the chrome::storage backend for the virtual
file system does not implement the storage::OpenFile API. However,
they all support the storage::FileStreamReader API.

Part 2 case: request the FuseBoxServer to read the file data using
storage::FileStreamReader and return the result.

The size_t |size| requested is saturated to int32_t on the client,
because the storage::FileStreamReader API `size` is int type.

BUG=chromium:1244007
TEST=emerge --board=nami fusebox && deploy

Change-Id: I5df230913a1c49f64508aacbc4ee33ad3d0b712b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3353069
Reviewed-by: François Degros <fdegros@chromium.org>
Commit-Queue: Noel Gordon <noel@chromium.org>
Tested-by: Noel Gordon <noel@chromium.org>