u2fd: Set max iterations for u2f_msg_handler_fuzzer
Currently we will use up all bytes in a single fuzzing round. However,
data size might be as large as millions of bytes, resulting in tens of
thousands of iterations, exceeding 90 seconds timeout for a single round
in ClusterFuzz. Set a max iteration (100) for it.

BUG=b:211493783
TEST=run testcase in the bug

Change-Id: I44ff5d72ffa1857365b5deb897a3a4608490e377
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3353642
Reviewed-by: Andrey Pronin <apronin@chromium.org>
Commit-Queue: Howard Yang <hcyang@google.com>
Tested-by: Howard Yang <hcyang@google.com>