cryptohome: Fix LowDiskSpaceHandler::Init sometimes not working
At present, the LowDiskSpaceHandler::Init() function sometimes fails
to run the disk cleanup routines successfully, because the thread that
runs the disk cleanup callbacks is different from the thread that calls
the Init function, and the Init function initialises an internal flag
too late in the process.

We need to mark the "stopped_" flag as false BEFORE calling any of the
callbacks, for the callbacks to work correctly; i.e. especially since
the default "base::TimeDelta()" is zero and the "post_delayed_task"
could call the callbacks from a different thread.

BUG=b:207825438
TEST="~/trunk/src/platform/tast-tests/tools/repeat-test.sh -s -r 100
      -- dut1-reverse cryptohome.AutomaticCleanupManyUsers.20_users"

Change-Id: I9b800b3ca9614f227a234d03782ada4a8d0552a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3353645
Tested-by: Mani Milani <mani@chromium.org>
Auto-Submit: Mani Milani <mani@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>
Commit-Queue: Andrey Pronin <apronin@chromium.org>