diagnostics: Add guards for status changes in NVMe routine
Currently in the NVMe routine, when the routine is cancelled between
GetNvmeLog and OnDebugdResultCallback, it leads to a race condition
causing the routine status overridden to running even if it’s cancelled.

Add guard conditions before updating the routine status to ensure the
modification only happens when the routine is not in the final state.

BUG=b:208438950
TEST=FEATURES=test USE="-cros-debug" emerge-${BOARD} diagnostics
TEST=tast run <IP> health.DiagnosticsRun.nvme_self_test
TEST=(w/o crrev/c/3308982 change) tast run <IP> \
                                  wilco.APIRoutine \
				  wilco.APIRoutineEnrolled

Change-Id: I782f532ec770a5ada5a98751c69ffa7f74964bed
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3356857
Tested-by: Denny Huang <dennyh@google.com>
Reviewed-by: Meng-Huan Yu <menghuan@chromium.org>
Commit-Queue: Denny Huang <dennyh@google.com>