ml: Retroactively fix stable mojom versioning
**This CL will need to be chumped** in order to bypass the stable mojom
presubmit.

Before there was a presubmit guarding it, some non-backwards-compatible
changes were made to [Stable] interfaces in this mojom. This CL
retroactively adds appropriate versioning information to make these
changes something that would have passed the presubmit (thus allowing
the changes to be synced to the Chrome side where the presubmit exists).

BUG=chromium:1081597, chromium:1283517
TEST=emerge-$B ml

Change-Id: I863295f1424ab5a18f1ef9e143f7742df62f7a6a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3360463
Tested-by: Andrew Moylan <amoylan@chromium.org>
Reviewed-by: Xinglong Luan <alanlxl@chromium.org>
Tested-by: Xinglong Luan <alanlxl@chromium.org>
Reviewed-by: Sam McNally <sammc@chromium.org>