vm_tools: Explicitly classify termina-like VMs in concierge
In several places we check the "is_termina" flag of the request to
determine if the request is for termina or some other VM. This will be
insufficient going forward as termina's infrastructure will be used for
most linux-like VMs. Instead, we explicitly classify the VM into a
concerge_service::VmInfo::VmType.

We do this because we need to distinguish borealis VMs for the purpose
of enforcing secure wayland connections. It is an error to launch a
borealis VM without such a wayland server.

BUG=b:200896773, b:207435650
TEST=attempt to launch borealis using VMC while it is disabled

Change-Id: Ia92e60d846f321cc9cb7a6be33cbccede447d20e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3360467
Reviewed-by: Fergus Dall <sidereal@google.com>
Tested-by: Chloe Pelling <cpelling@google.com>
Commit-Queue: Chloe Pelling <cpelling@google.com>