hps: send ERASE_STAGE1 command before writing stage1
In https://crrev.com/c/3349853 we introduced an explicit command to
erase the RW region of MCU flash. Send that command before writing
stage1.

If this is run against older stage0, where that command is not
implemented, it will fault with I2CBADREQ but the subsequent writes will
succeed as normal and the fault will be cleared on the following reset.
So this also works correctly if stage0 is not updated.

BUG=b:210734629
TEST=(brya, WP on) truncate -s 65535 empty64k.bin &&
     hps dl 0 empty64k.bin && start hpsd  # HPS boots, stage1 verified

Change-Id: Ic07f6fdabe3abedeb27d0288f0a7fd856b9f7e16
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3367576
Tested-by: Dan Callaghan <dcallagh@chromium.org>
Reviewed-by: David Lattimore <dml@chromium.org>
Commit-Queue: Dan Callaghan <dcallagh@chromium.org>