chaps: Correctly handle null value in attributes
By design of the format of attribute, we can have an attribute with
pValue is nullptr, but length is defined. When receiving this kind of
attribute in SetAttributes, currently it will try read up to `length`
bytes of data from `pValue` which is null, resulting in a crash. Fix it
to correctly handle null value.

BUG=chromium:172219720
TEST=chaps unittest

Change-Id: Iacc160b3783f3145f5ccb3952d4a35237ce9ef4f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3367585
Reviewed-by: John L Chen <zuan@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>
Tested-by: Howard Yang <hcyang@google.com>
Commit-Queue: Howard Yang <hcyang@google.com>