shill: Report cellular techs in Connect Result metric
To track connection success rates across different technologies, this
patch adds the technologies in use to connection result metric.

We report the raw access technologies bitmask because it is possible for
more than one technology to be in use at once, for instance in 5G-NSA.
This granularity will allow us to compare the technologies properly.

BUG=b:186800810
TEST=emerge'd metrics and shill on brya, validated new field in logs

Change-Id: I8b18d00bc28054a46ec8fc43d0feb3d450c416a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3371044
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Tony Yeoman <tby@chromium.org>
Tested-by: Daniel Winkler <danielwinkler@google.com>
Commit-Queue: Daniel Winkler <danielwinkler@google.com>