vm_tools: Use midpoint in balloon's balance available
When calculating the delta in the balloon's balance available policy,
split the difference between the host and guest available values. This
accounts for the fact that inflating the balloon to increase host
available decreases guest available by a corresponding amount. Deflating
the balloon to increase guest available doesn't necessarily lead to a
1-to-1 decrease in host available, due to on-demand allocation of pages
and zram, but the principle is the same.

BUG=b:213962590
TEST=verify no balloon oscillation with crrev.com/c/3394436

Change-Id: Ic6c2eb2caa9c5902e28481dd593f25ec494df80e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3394428
Reviewed-by: Charles William Dick <cwd@google.com>
Reviewed-by: Abhishek Bhardwaj <abhishekbh@chromium.org>
Tested-by: David Stevens <stevensd@chromium.org>
Commit-Queue: David Stevens <stevensd@chromium.org>