runtime_probe: Capture logging of helper process
To enhance the development experience and to reduce the debugging
effort, the stderr of the helper process is captured and can be logged
in the caller.
With this, we don't need to log to syslog in the helper, thus we can
omit the syscall necessary for syslog in seccomp. Now the helper
always log to stderr.

This change also prevents executing |EvalImpl()| in |EvalInHelper()| of
non privileged probe function.

BUG=b:185292404
TEST=emerge-zork debugd debugd-client runtime_probe
TEST=test on dut

Change-Id: I73b4c75597b621bd5e0489095e20628b47775473
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3372388
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Chung-Sheng Wu <chungsheng@chromium.org>
Reviewed-by: Kevin Lin <kevinptt@chromium.org>
Commit-Queue: Chung-Sheng Wu <chungsheng@chromium.org>