vm_tools: cros_im: Support control characters like tab and enter
Fix keysym support to set the hardware keycode appropriately for control
characters. This fixes for example pressing enter to activate a linked
button on a single line text entry, or pressing tab to change focus.
Unlike wl_keyboard::key, the flow here is a bit roundabout as we don't
have the hardware keycode here. The current approach appears to work but
if issues are identified later we may need an API which directly sends
the hardware keycodes.

BUG=chromium:1247564
TEST=build on DUT as per README.md, ../test/run_tests.py
TEST=run gedit with cros_im and save a file by pressing enter when focus
is on the file name text entry.

Change-Id: Ibce176cc9c8718cd8915c9dccd45cf90a64204c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3375427
Reviewed-by: David Munro <davidmunro@google.com>
Commit-Queue: Timothy Loh <timloh@chromium.org>
Tested-by: Timothy Loh <timloh@chromium.org>