shill: passpoint: fill services on matches.
On Passpoint credentials match, the provider receives the set of
matches. The provider keeps the best possible match for each service and
fill it with the EAP credentials, a match score and a reference to the
Passpoint credentials. The service is finally marked as connectable and
will be able to authenticate using EAP-{TLS,TTLS}.

The match score is computed from the properties provided during a match.
It represents a priority level common to all the services who've been
populated through a Passpoint match and helps selecting the best one
when multiple are connectable. For now it is directly computed from the
network type (home, roaming, ...) but it might be complexified by the
introduction of credentials priorities later.

The reference to the parent Passpoint credentials is kept to provide the
"metered override" information, and will allow later to remove all the
services affected by one set of credentials.

BUG=b:162106001
TEST=unit test;
Push a set of credentials to Shill using 'gdbus call' and create
successively multiple matching access points. Check it connects
automatically and successfully to each of them. Reboot and ensure the
networks are still present in the "Wi-Fi > Known networks" menu.

Change-Id: I6969ff09ce0ce941972a98b5a93af13aa7781c83
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3378628
Tested-by: Damien Dejean <damiendejean@google.com>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Auto-Submit: Damien Dejean <damiendejean@google.com>
Commit-Queue: Damien Dejean <damiendejean@google.com>