fusebox: DirEntryRequest: call libFUSE fuse_add_direntry once
If there is not enough space in the output |buf| to add an entry,
fuse_add_direntry() does not add the entry. But its return is the
amount of buffer space `used` by the entry, whether added or not.

If this entry was not added then |used| will exceed the available
buffer |size|. Use that to figure out if the entry was added.

Result: only need to call fuse_add_direntry() once, and no change
in behavior.

BUG=chromium:1244007
TEST=emerge-${BOARD} fusebox

Change-Id: I9bde9ace18a65abd32efa74cd439dea67de81c01
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3379589
Commit-Queue: François Degros <fdegros@chromium.org>
Reviewed-by: François Degros <fdegros@chromium.org>
Commit-Queue: Noel Gordon <noel@chromium.org>
Tested-by: Noel Gordon <noel@chromium.org>