fusebox: create the DirEntryResponse in FUSE op ReadDir
Doing so simplifies the ReadDirResponse code flow: the server always
replies OK to a `readdir` request. So if the server doesn't reply it
is an error (the server has gone, crashed, etc). Append the error to
the handle's DirEntryResponse. GetResponseErrno() logs the error.

The `readdir` protocol is now robust w.r.t. server error, and to the
order of server replies to a `readdir` request [1]. The OK reply and
DirEntry stream replies can arrive in any order: the client does not
depend on reply order.

BUG=chromium:1244007
TEST=emerge-${BOARD} fusebox

Change-Id: Ia3ad9183639e220b114ad26cf2c8ef99028977c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3379885
Reviewed-by: Ben Reich <benreich@chromium.org>
Commit-Queue: Noel Gordon <noel@chromium.org>
Tested-by: Noel Gordon <noel@chromium.org>