shill: drop supplicant interface get/create/remove errors to warnings
The wireless interface state sometimes races with shill. Shill's
inability to get/create/remove an interface in wpa_supplicant is not
necessarily an error. Drop these to warnings knowing that a) we'll log
an error if we fail these operations too many times and b) we'll log an
error in wpa_supplicant or from the kernel if something is actually
wrong.

BUG=b:213629562
TEST=`emerge-${BOARD} shill`

Change-Id: Ia1c05e2901d219961e295ce381663ade47479a6e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3380204
Reviewed-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Tested-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Matthew Wang <matthewmwang@chromium.org>