shill: passpoint: credentials removal.
When a service populated using Passpoint is removed, the parent set of
credentials and all the linked services have to be removed. Thus, on
Wi-Fi service removal, the provider gets the Passpoint credentials
through OnServiceUnloaded() if any. Then the provider removes the set of
credentials from the device, from its list and removes all the Wi-Fi
services that were using it.

BUG=b:162106001
TEST=unit tests;
Push a set of credentials to Shill using "gdbus call", create a
compatible access point, let the device match and connect. Repeat the
operation with multiple access points and the same set of credentials.
From the Wi-Fi menu, remove one of the networks. Check all the networks
obtained from Passpoint credentials are removed, check the device is not
able to associate automatically with a Passpoint access point.

Change-Id: I80a02b7329c62def732afbd11895d00b5c1ae2ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3378629
Tested-by: Damien Dejean <damiendejean@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Auto-Submit: Damien Dejean <damiendejean@google.com>
Commit-Queue: Damien Dejean <damiendejean@google.com>