hiberman: Wire together cryptohome and Chrome dbus calls
Create the ResumeFromHibernate interface that Chrome can call in
to initiate resume. This lets the hiberman service know that
authentication has completed and the seed is available for fetching from
cryptohome. With this, we can actually wire up the whole end to end
resume flow.

There were also some locking issues, in that the
HiberDbusConnectionInternal structure is locked for as long as the dbus
thread is running, but the main thread was trying to access
HibernateDbusStateInternal via that structure. You cannot both do this
and have the dbus thread block in the method until the main thread
releases it. So instead, lift out the HibernateDbusStateInternal so that
it can be refcounted and locked directly from HiberDbusConnection,
without having to acquire the internal connection lock. Also return a
PendingResumeCall, which wraps the dbus connection and represents a
pending call to release the blocked method when dropped.

BUG=b:204563517
TEST=hibernate on volteer

Change-Id: I1b83e0626c7e03a78279a5bdac446aa5b1e3257d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3380248
Tested-by: Evan Green <evgreen@chromium.org>
Reviewed-by: Allen Webb <allenwebb@google.com>
Commit-Queue: Evan Green <evgreen@chromium.org>