shill: vpn: Enable swanctl-based driver by default
Except for the cases where tunnel group is set.
Note that there is a flaw that the driver is decided when
creating/reloading the service, so only changing the field in the UI
will not trigger the switch, a reboot (or restarting shill) is needed.

BUG=b:204261554
TEST=manually verified that the new driver is used by default, and then
  input something in the tunnel group field, restarted shill, verified
  that the legacy driver is used (the connection failed at auth stage
  though).

Change-Id: Ida0625f5bde5623ec48935a9f4f9b2db08b04952
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3388897
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>