sirenia: add starting/stopping VMs to MMS
Integrate starting and stopping VMs with ManaTEE memory service, and
keep track of running VMs.

Since the hypervisor cannot handle memory overcommit, memory needs to be
set aside for a new VM before the VM can be launched. This means that
starting a VM is a three step process from the point of view of the CrOS
guest: first, it asks MMS to set aside memory, then it starts the VM,
and finally it tell MMS that the VM was started. In addition to
supporting this normal path, MMS needs to support the various failure
cases caused by errors or crashes in the CrOS guest client process.

Compared to startup, stopping VMs is straightforward. Assuming that the
CrOS guest has triggered shutdown in the target VM, MMS simply needs to
wait for the target VM to exit and then release any memory allocated by
that VM back to the CrOS guest.

BUG=b:213962590
TEST=manual testing with crrev.com/c/3394436

Change-Id: I57a9a0016a41fe0d4335bb470d3864ce99e49ffd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3394434
Reviewed-by: Allen Webb <allenwebb@google.com>
Tested-by: David Stevens <stevensd@chromium.org>
Commit-Queue: David Stevens <stevensd@chromium.org>