modemfwd: reject absolute paths in manifest v1
These cause a DCHECK in base::FilePath::Append
and generally should not be present in the manifest
to begin with. So, reject the manifest as malformed
if we see one.

This check already exists in the parser for manifest
V2, but we can add a unit test for it as well.

BUG=b:198327637
TEST=unit tests

Change-Id: Ie6e849e58520cd7c1223e4e74937d075e06a7b37
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3388531
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Auto-Submit: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>
Commit-Queue: Eric Caruso <ejcaruso@chromium.org>