fusebox: add MakeTimeStat() helper
Add a helper to create a stat with mode and time fields set to a
given |mode| and |time| (default: the current time).

Add a test for the helper and update the file system fake to use
it when creating the fake file system.

BUG=chromium:1240983
TEST=cros_workon_make fusebox --test

Change-Id: If3db3075c32f76133e80e32000de5b9ddfb04d1f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3393224
Reviewed-by: François Degros <fdegros@chromium.org>
Tested-by: François Degros <fdegros@chromium.org>
Commit-Queue: Noel Gordon <noel@chromium.org>
Tested-by: Noel Gordon <noel@chromium.org>