arc-setup: Remounting data directory as NOEXEC
In the init namespace, we have a scenario where
/home/root/<hash>/android-data/data is both RW and EXEC.
This is a vulnerability. Ultimately, ARC container
needs the EXEC capability, but we want the above
directory to be NO_EXEC only.

The above mount point is first created, passed to ARC and
then unmounted. Long time window increases the vulnerability.
This CL reduces that window by remounting it as NO_EXEC
as soon after the shared mount is passed without waiting
until is unmounted altogether.

BUG=b:213625515
TEST=arc.Boot
TEST=Ensure that /data directory inside the ARC container
has EXEC capability.

Change-Id: Ief71ae283fe9fc446dcdc00369a77aa05c44a5cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3396819
Reviewed-by: Yusuke Sato <yusukes@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Commit-Queue: Vaibhav Raheja <vraheja@chromium.org>
Tested-by: Vaibhav Raheja <vraheja@chromium.org>