chromeos-config: Add branding info to schema
Marketing & OEM names are now under `arc/build-properties/` in
cros-config, but it makes more sense to find another place to store that
information.

A top-level field `branding` is created to organize these fields. In
addition, to prevent leaking the OEM name in the development stage, a
warning is also added to the description.

BUG=b:210972707
TEST=./regen.sh && ./run_tests.sh (lint errors ignored)

Cq-Depend: chromium:3385463
Change-Id: I27addb2ba3da773e0df51478625e9dd0ab748e40
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3397879
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Meng-Huan Yu <menghuan@chromium.org>
Reviewed-by: Sam McNally <sammc@chromium.org>
Tested-by: Denny Huang <dennyh@google.com>
Commit-Queue: Denny Huang <dennyh@google.com>