login: Write/delete Chromad migration file
Add logic for writing kChromadMigrationSkipOobePreservePath file when a
remote powerwash request is initiated in AD mode, as part of the Chromad
migration. This flag file indicates that some OOBE screens should be
skipped after the device is powerwashed.

Also, add logic for deleting this file, when we detect that the device
has already been re-enrolled in cloud management mode.

Additionally, this file is added to the list of powerwash safe files in
ClobberState, been kept during any safe wipe.

BUG=chromium:1209246
TEST=cros_run_unit_tests --board=${BOARD} \
     --packages "chromeos-login chromeos-init"

Change-Id: I3bd59598328bb82db44e9e4fd2f919a977791441
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3399841
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Roman Sorokin <rsorokin@chromium.org>
Reviewed-by: Mattias Nissler <mnissler@chromium.org>
Tested-by: Felipe Andrade <fsandrade@chromium.org>
Commit-Queue: Felipe Andrade <fsandrade@chromium.org>