fusebox: remove |ptr| name from the device attachment API
The methods are auto-generated by the DBUS bindings system, so we
have no control over the API signature [1].

We do not use |ptr|, and never need to: the success or failure of
the methods is sent back to the FuseBoxServer in the bool return,
and any errors LOG to syslog. Remove the |ptr| parameter name.

[1] Other parts of the FuseBox code could call the API, to attach
or detach during tests, or during startup, for example. The auto-
generated DBUS code is fine and all, but also inflexible if other
parts of the code that are not DBUS want to call these methods.

BUG=chromium:1289493
TEST=emerge-${BOARD} fusebox

Change-Id: I619b40668798439f9434148df2201cda47faf910
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3406235
Reviewed-by: Jeremie Boulic <jboulic@chromium.org>
Commit-Queue: Noel Gordon <noel@chromium.org>
Tested-by: Noel Gordon <noel@chromium.org>