vm_tools: cros_im: Fix handling of sub-windows
The "Open" button in gedit opens a GTK popover with an Entry that gets
focus, but currently text input does not work there with the cros IM
module. It seems like for keysym events we need to get the effective top
level window for both the GTK and GDK windows to make this work.

BUG=chromium:1247564
TEST=ninja && ../test/run_tests.py
TEST=Run gedit, click "Open", input text. Also use IME on various text
fields and check IME UI is positioned correctly next to the cursor.

Change-Id: Id225050e7647fce7335d5cef4f9e3c4c1f6167cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3412588
Tested-by: Timothy Loh <timloh@chromium.org>
Reviewed-by: David Munro <davidmunro@google.com>
Commit-Queue: Timothy Loh <timloh@chromium.org>