login: Add UMA stats for Chrome state when killed
Add a UMA histogram recording the state of the Chrome browser process at
the moment the liveness checker decides the process is hung. (Normally,
the moment it sends the abort signal, if enable_aborting_ is true.)

We can sometimes get this information from the "top" dump in the logs,
but the free-text nature of the logs makes analysis-at-scale difficult.
This will give us better insights into why Chrome is not responding to
liveness pings when it is not responding to liveness pings.

histograms.xml updates in
https://chromium-review.googlesource.com/c/chromium/src/+/3418458

BUG=b:214076950
TEST=Forced timeout to 1 second, viewed chrome://histograms, saw record
with value 1 (Sleeping)
TEST=Suspended Chrome with kill -19, saw record in
/var/lib/metrics/uma-events with value 4 (TracedOrStopped)

Change-Id: I39d59e7f12c94d30ae97c8af0fdadfe169ac1bef
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3418877
Commit-Queue: Ian Barkley-Yeung <iby@chromium.org>
Tested-by: Ian Barkley-Yeung <iby@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>