fusebox: move a C++ cast into low-layer FUSE ReplyBuffer
Make ReplyBuffer accept const void* and perform the cast inside FUSE
low-layer ReplyBuffer code. No change in behavior.

There are other C++ and base:: casts in class FuseBoxClient and they
all exist for security reasons.

The reinterpret_cast<> moved in this CL was not for security: it was
only used to convert a D-BUS type to a libFUSE type.

BUG=chromium:1240983
TEST=emerge-${BOARD} fusebox

Change-Id: I25ce943998f05e0d048fae17757df6d7b4a5aa71
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3412576
Reviewed-by: Ben Reich <benreich@chromium.org>
Commit-Queue: Noel Gordon <noel@chromium.org>
Tested-by: Noel Gordon <noel@chromium.org>