trunks: Stabilize the output of generator
The result of traversing a set in Python is not stable.

The unstable result would lead to different generated result and makes
the generated files hard to diff between multiple version.

To make the output of generator more stable, we should returning a
list instead of a set for the traversing purpose.

BUG=b:212660790
TEST=Build OK.
TEST=generate 10 times, and the output is stable.

Change-Id: Ie35a2a3d5a68aaa2235aecd02ffd08ac833bc4bb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3412590
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>
Commit-Queue: Yi Chou <yich@google.com>
Tested-by: Yi Chou <yich@google.com>