login: Add metrics around arc-boot-continue impulse
Currently we are only logging the time taken for arc-boot-continue
impulse time when its successful. However to identify cases when the
operation times out or is unsuccessful we need a metric. We also need to
get an accurate timeout value for the same.

In this change we are adding a new metrics as well as increasing the
buckets for exiting metric to get a more granular timeout value we
should set in future.

histograms.xml updates in
https://chromium-review.googlesource.com/c/chromium/src/+/3430517

BUG=b:205032502
TEST=tast run -build=false DUT arc.Optin
TEST=cros_run_unit_tests --board=${BOARD} --packages chromeos-login

Change-Id: I8b6e8c8fd03f18558ba23a186a977c963cb7f6d3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3414910
Reviewed-by: Yusuke Sato <yusukes@chromium.org>
Reviewed-by: Muhammad Hasan Khan <mhasank@chromium.org>
Tested-by: Varun Somani <vsomani@google.com>
Auto-Submit: Varun Somani <vsomani@google.com>
Commit-Queue: Varun Somani <vsomani@google.com>