missive: Ignore Delete().ok in RemoveConfirmedData
Sync up to components: crrev.com/c/3416130

Currently if the file fails to be deleted for some reason, we'll get
stuck in the loop. Ignore the deletion result as it is not a big problem
to leave the file on disk in case of such an unusual failure.

FIXED=b:215637295
BUG=b:215637295
TEST=FEATURES="test" emerge-$BOARD missive

Change-Id: I5305764e99d9094a29be96d27b65c3cba3708ab3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3416095
Reviewed-by: Leonid Baraz <lbaraz@chromium.org>
Commit-Queue: Hong Xu <xuhong@google.com>
Tested-by: Hong Xu <xuhong@google.com>
Auto-Submit: Hong Xu <xuhong@google.com>