cryptohome: [USS] Get wrapping key from authblocks
Obtain the per-credential secret for UserSecretStash by applying HKDF
to the auth block's output. This adds a missing link between the auth
blocks component (which so far has been vault keyset oriented) and
the USS component (which needs a per-credential secret to wrap the
USS Main Key). The usage of HKDF for that link, as opposed to directly
reusing the existing "vkk_key" value, avoids the issue of key reuse
between different algorithms, without adding new complexity into auth
block implementations.

Follow-ups are planned to rename the "vkk_key" field to a more generic
name, reflecting the fact it'll be used not only for vault keysets,
and to potentially hide it as a private field that's only used for
derivation of returned secrets.

BUG=b:216474361
TEST=cros_workon_make --board=$BOARD --test cryptohome

Change-Id: I20840330192051c6508a204fcef36a2f59525bd5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3416995
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>
Reviewed-by: Greg Kerr <kerrnel@chromium.org>
Commit-Queue: Greg Kerr <kerrnel@chromium.org>
Tested-by: Greg Kerr <kerrnel@chromium.org>