secanomalyd: Report mounts and processes.
Crash reports for anomalous system states are not very useful if they
only contain the actual anomaly but not a description of the system.

Report the full list of mounts (excluding USB drives and archive
files), and the list of processes in the init PID namespace to get an
idea of what's going on in anomalous systems.

USB drives and archive mounts are excluded because they can contain
PII, processes not in the init PID namespace are excluded because they
too can contain PII.

In order to implement this we introduce the concept of a "Filter" which
allows the information-gathering functions to easily select which
mounts or processes to include in the report.

Process reporting is fully implemented but not yet hooked up because it
requires an extra capability. Enabling that will be done in a follow-up
CL to avoid changing the runtime permissions of the daemon in a big CL
like this one.

BUG=b:216855846
TEST=New unit tests.
TEST=Deploy, run, check crash report directory at /var/spool/crash.

Change-Id: Iefb79ce608f834fcd66e4af9c67f5de9c640eb6d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3424666
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Auto-Submit: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Aashay Shringarpure <aashay@google.com>
Commit-Queue: Jorge Lucangeli Obes <jorgelo@chromium.org>