cryptohome: [USS] Helper for main key generation
Add a method for generating the USS main key into the UserSecretStash
class. It's a thin wrapper around the crypto primitive, but it's still
useful as code outside the USS layer shouldn't need to know the type
and the size of the key.

BUG=b:216736218
TEST=cros_workon_make --board=$BOARD --test cryptohome

Change-Id: I645be9a2699370ad12c6f1cc7528e5b1923556ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3422629
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>
Auto-Submit: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Greg Kerr <kerrnel@chromium.org>
Commit-Queue: Greg Kerr <kerrnel@chromium.org>
Tested-by: Greg Kerr <kerrnel@chromium.org>