shill: technology specific order for services.
Services are sorted with the Service::Compare() function using the
fields available in the Service class. To avoid exposing technology
specific criteria in the base Service class, this CL brings a
CompareWithSameTechnology() method that can be implemented per
technology to perform a comparison. It's designed to be called with two
services of the same technology and the condition is guaranteed by
Service::Compare().

The first technology specific method is implemented in WiFiService and
adds Passpoint network ordering criteria.

BUG=b:162106001
TEST=unit test

Change-Id: I0ee109c219645398382303452b05146bd485ac61
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3423788
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Tested-by: Damien Dejean <damiendejean@google.com>
Reviewed-by: Pavol Marko <pmarko@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Auto-Submit: Damien Dejean <damiendejean@google.com>
Commit-Queue: Damien Dejean <damiendejean@google.com>