shill: change SetProperty to be void
Previously, SetProperty returned a boolean that was only used in tests.
Changed all instances in tests to instead check the error object for
Success or Failure states of setting properties. SetProperty and the
other SetProperty functions for different variables now return void.

Along with this, changed checks for equivalence of error type and
kSuccess to instead use IsSuccess(), since that function does the
same thing and is a little more clear.

BUG=b:183624193
TEST=FEATURES="test" emerge-${BOARD} shill

Change-Id: I41bdb7d9df26e4df88d7cd159e4dc4ad192ee95a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3425082
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Tested-by: Michael Rygiel <michaelrygiel@google.com>
Commit-Queue: Michael Rygiel <michaelrygiel@google.com>