shill: Add connection attempt/result metrics
When we attempt to connect to an AP, emit the WiFiConnectionAttempt
event that includes information about the type of security settings, the
channel/band, salted+hashed (B)SSID, etc.
Also add some more information about relevant features supported by the
AP, such as Passpoint, MBO, etc.

When the connection attempt succeeds (or fails...), emit the
WiFiConnectionAttemptResult event to know if the connection attempt was
successful or not.

BUG=b:203692510
TEST=cros-workon-${BOARD} start metrics shill
TEST=emerge-${BOARD} metrics shill
TEST=FEATURES=test emerge-${BOARD} shill
TEST=cros deploy ${DUT} metrics shill
TEST=restart shill
TEST=Observe in the log with extra debug messages that the we log
connection attempts and results. Test failure to connect by mistyping
the password on purpose, observe in the log that the event is generated.

Cq-Depend: chromium:3448157
Change-Id: I2578393d89ddac3f88ffabf89675959f59be3e26
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3428949
Tested-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Tony Yeoman <tby@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Nicolas Norvez <norvez@chromium.org>